requirements.txt
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
